refactor: сorrections based on comments pr 15
