+++ src/modules/expenses/enum/ExpensesFilterEnum.ts
@@ -1,7 +1,7 @@
 export enum ExpensesFilterEnum {
-    CUSTOMER_ID = 'customerId',
-    CONTRACT_NUMBER = 'contractNumber',
-    ASSET = 'asset',
+    CLIENT_ID = 'clientId',
+    ASSET_ID = 'assetId',
+    CLIENT_CONTRACT_ID = 'clientContractId',
     START_DATE = 'startDate',
     END_DATE = "endDate",
     STATUS = 'status',
