+++ src/modules/user/hooks/useAuthLogOut.ts
@@ -1,6 +1,6 @@
 import { useDispatch, useSelector } from 'react-redux';
 import { useNavigate } from 'react-router-dom';
-import { useLogoutMutation } from '../services/apiService';
+import { useLogoutMutation } from '../../../app/services/apiService';
 import { RoutesEnum } from '../../../app/routes/RoutesEnum';
 import { logOut as logoutAction } from '../slice/authSlice';
 import { setShowLogout, selectLogOut } from '../slice/authSlice';
